requirements.txt
torch
torchvision
transformers
pandas
scikit-learn
tqdm
Pillow
PyYAML
nltk
rouge_score
